# docs/swagger.yaml
definitions:
  request.AddUser:
    properties:
      email:
        type: string
      mobile:
        type: string
      nickname:
        maxLength: 16
        minLength: 2
        type: string
      password:
        maxLength: 16
        minLength: 6
        type: string
      status:
        type: integer
      username:
        maxLength: 16
        minLength: 2
        type: string
    required:
    - nickname
    - password
    - status
    - username
    type: object
...
  request.AssignRole:
    properties:
      role_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  request.ChangePassword:
    properties:
      confirm_password:
        maxLength: 16
        minLength: 6
        type: string
      old_password:
        maxLength: 16
        minLength: 6
        type: string
      password:
        maxLength: 16
        minLength: 6
        type: string
    required:
    - confirm_password
    - old_password
    - password
    type: object
...
  request.Login:
    properties:
      password:
        maxLength: 16
        minLength: 6
        type: string
      username:
        maxLength: 16
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  request.UpdateUser:
    properties:
      email:
        type: string
      mobile:
        type: string
      nickname:
        maxLength: 16
        minLength: 2
        type: string
      password:
        type: string
      status:
        type: integer
      username:
        maxLength: 16
        minLength: 2
        type: string
    required:
    - nickname
    - username
    type: object
  request.UserList:
    properties:
      currentPage:
        type: integer
      pageSize:
        type: integer
      status:
        type: string
      username:
        type: string
    type: object
host: 127.0.0.1:9111
info:
  contact: {}
  description: gin框架后台管理
  title: gin框架后台管理demo
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      parameters:
      - description: 用户名,密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"token":"token","expiresAt":1656042399},"msg":"登录成功！"}'
          schema:
            type: string
      summary: 登录
      tags:
      - Base
  /api/v1/auth/loginOut:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"退出成功！"}'
          schema:
            type: string
      summary: 退出
      tags:
      - User
...
  /api/v1/user/changepassword:
    put:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 旧密码,新密码,确认密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"修改密码成功！"}'
          schema:
            type: string
      summary: 修改密码
      tags:
      - User
  /api/v1/user/getRoleIdByUserId/:id:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id, 角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AssignRole'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"分配角色成功！"}'
          schema:
            type: string
      summary: 分配角色
      tags:
      - User
  /api/v1/user/getRouteMenuList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"route_tree":[]}]}]},"msg":"获得Route菜单成功！"}'
          schema:
            type: string
      summary: 用户Route菜单
      tags:
      - User
  /api/v1/user/getUserList:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserList'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"list":[],"total":4,"current_page":1,"page_size":20},"msg":"获取用户列表成功！"}'
          schema:
            type: string
      summary: 系统用户列表
      tags:
      - User
  /api/v1/user/info:
    get:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{"id":1,"uuid":"uuid","username":"username","nickname":"nickname","avatar":"avatar","status":1,"mobile":"","email":"","roles":[]},"msg":"获得登录用户信息成功！"}'
          schema:
            type: string
      summary: 登录用户基本信息
      tags:
      - User
  /api/v1/user/user:
    post:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名,密码,昵称,状态
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AddUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"添加用户成功！"}'
          schema:
            type: string
      summary: 添加系统用户
      tags:
      - User
  /api/v1/user/user/:id:
    delete:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"删除用户成功！"}'
          schema:
            type: string
      summary: 删除用户
      tags:
      - User
    put:
      parameters:
      - description: 验证参数Bearer和token空格拼接
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名,昵称
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":null,"msg":"编辑用户成功！"}'
          schema:
            type: string
      summary: 编辑用户
      tags:
      - User
swagger: "2.0"